requirements.txt
# Application de Recommandation Agricole - Django Web App

# Framework Web
Django>=5.2.0
django-environ>=0.11.0

# Machine Learning
scikit-learn>=1.4.0
pandas>=2.1.0
numpy>=1.26.0

# Base de données
PyMySQL>=1.1.0
mysql-connector-python>=8.2.0
SQLAlchemy>=2.0.23

# API et configuration
requests>=2.31.0
python-dotenv>=1.0.0